PLT-7287 Move access token text in user list to under the dropdown (#7099)
* Move access token text in user list to under the dropdown

* Updating UI for token help text (#7107)
