+++ webapp/i18n/en.json
@@ -988,11 +988,9 @@
   "admin.manage_roles.manageRolesTitle": "Manage Roles",
   "admin.manage_roles.cancel": "Cancel",
   "admin.manage_roles.save": "Save",
-  "admin.user_item.userAccessTokenNo": "<strong>User Access Tokens:</strong> No",
-  "admin.user_item.userAccessTokenAdmin": "<strong>User Access Tokens:</strong> Yes (with system_admin)",
-  "admin.user_item.userAccessTokenPostAll": "<strong>User Access Tokens:</strong> Yes (with post:all)",
-  "admin.user_item.userAccessTokenPostAllPublic": "<strong>User Access Tokens:</strong> Yes (with post:channels)",
-  "admin.user_item.userAccessTokenYes": "<strong>User Access Tokens:</strong> Yes",
+  "admin.user_item.userAccessTokenPostAll": "(with post:all user access tokens)",
+  "admin.user_item.userAccessTokenPostAllPublic": "(with post:channels user access tokens)",
+  "admin.user_item.userAccessTokenYes": "(with user access tokens)",
   "admin.user_item.authServiceEmail": "<strong>Sign-in Method:</strong> Email",
   "admin.user_item.authServiceNotEmail": "<strong>Sign-in Method:</strong> {service}",
   "admin.user_item.confirmDemoteDescription": "If you demote yourself from the System Admin role and there is not another user with System Admin privileges, you'll need to re-assign a System Admin by accessing the Mattermost server through a terminal and running the following command.",
